Splitting CsDebugScript.Tests into Tests.Common, Tests.Native and Tests.CLR
