+++ Source/CsDebugScript.DbgEng/Properties/AssemblyInfo.cs
@@ -1,5 +1,5 @@
 ﻿using System.Runtime.CompilerServices;
 
 // Allow internals visibility to other projects
-[assembly: InternalsVisibleTo("CsDebugScript.Tests")]
+[assembly: InternalsVisibleTo("CsDebugScript.Tests.Native")]
 [assembly: InternalsVisibleTo("CsDebugScript.Scripting")]
