+++ Source/CsDebugScript.PdbSymbolProvider/Properties/AssemblyInfo.cs
@@ -1,4 +1,4 @@
 ﻿using System.Runtime.CompilerServices;
 
 // Allow internals visibility to other projects
-[assembly: InternalsVisibleTo("CsDebugScript.Tests")]
+[assembly: InternalsVisibleTo("CsDebugScript.Tests.Native")]
