+++ Source/CsDebugScript.Scripting/Properties/AssemblyInfo.cs
@@ -3,4 +3,5 @@
 // Allow internals visibility to other projects
 [assembly: InternalsVisibleTo("CsDebugScript.UI")]
 [assembly: InternalsVisibleTo("CsDebugScript.VS")]
-[assembly: InternalsVisibleTo("CsDebugScript.Tests")]
+[assembly: InternalsVisibleTo("CsDebugScript.Tests.Common")]
+[assembly: InternalsVisibleTo("CsDebugScript.Tests.Native")]
