+++ Tests/CsDebugScript.Tests.CLR/AppDomainTests.cs
