+++ Tests/CsDebugScript.Tests.CLR/Apps/AppDomains.cs
