+++ Tests/CsDebugScript.Tests.CLR/Apps/LocalVariables.cs
