+++ Tests/CsDebugScript.Tests.CLR/Apps/NestedException.cs
