+++ Tests/CsDebugScript.Tests.CLR/ClrDumpInitialization.cs
