+++ Tests/CsDebugScript.Tests.CLR/ExceptionTests.cs
