+++ Tests/CsDebugScript.Tests.CLR/ModuleTests.cs
