+++ Tests/CsDebugScript.Tests.CLR/Properties/AssemblyInfo.cs
