+++ Tests/CsDebugScript.Tests.CLR/StackTests.cs
