+++ Tests/CsDebugScript.Tests.CLR/ValueTests.cs
