+++ Tests/CsDebugScript.Tests.Common/Properties/AssemblyInfo.cs
