+++ Tests/CsDebugScript.Tests.Common/SkippableFactAttribute.cs
@@ -1,9 +1,9 @@
 ﻿using Xunit;
 using Xunit.Sdk;
 
-namespace CsDebugScript.Tests.Utils
+namespace CsDebugScript.Tests
 {
-    [XunitTestCaseDiscoverer("CsDebugScript.Tests.Utils.SkippableFactDiscoverer", "CsDebugScript.Tests")]
+    [XunitTestCaseDiscoverer("CsDebugScript.Tests.SkippableFactDiscoverer", "CsDebugScript.Tests")]
     public class SkippableFactAttribute : FactAttribute
     {
         public string SkipOnFailurePropertyName { get; set; }
