+++ Tests/CsDebugScript.Tests.Common/SkippableTheoryAttribute.cs
@@ -1,0 +1,10 @@
+﻿using Xunit;
+using Xunit.Sdk;
+
+namespace CsDebugScript.Tests
+{
+    [XunitTestCaseDiscoverer("CsDebugScript.Tests.SkippableTheoryDiscoverer", "CsDebugScript.Tests")]
+    public class SkippableTheoryAttribute : TheoryAttribute
+    {
+    }
+}
