+++ Tests/CsDebugScript.Tests.Common/SkippableTheoryTestCase.cs
@@ -4,15 +4,15 @@ using System.Threading.Tasks;
 using Xunit.Abstractions;
 using Xunit.Sdk;
 
-namespace CsDebugScript.Tests.Utils
+namespace CsDebugScript.Tests
 {
-    public class SkippableFactTestCase : XunitTestCase
+    public class SkippableTheoryTestCase : XunitTheoryTestCase
     {
         [Obsolete("Called by the de-serializer; should only be called by deriving classes for de-serialization purposes")]
-        public SkippableFactTestCase() { }
+        public SkippableTheoryTestCase() { }
 
-        public SkippableFactTestCase(IMessageSink diagnosticMessageSink, TestMethodDisplay defaultMethodDisplay, ITestMethod testMethod, object[] testMethodArguments = null)
-            : base(diagnosticMessageSink, defaultMethodDisplay, testMethod, testMethodArguments) { }
+        public SkippableTheoryTestCase(IMessageSink diagnosticMessageSink, TestMethodDisplay defaultMethodDisplay, ITestMethod testMethod)
+            : base(diagnosticMessageSink, defaultMethodDisplay, testMethod) { }
 
         public override async Task<RunSummary> RunAsync(IMessageSink diagnosticMessageSink,
                                                         IMessageBus messageBus,
@@ -20,6 +20,7 @@ namespace CsDebugScript.Tests.Utils
                                                         ExceptionAggregator aggregator,
                                                         CancellationTokenSource cancellationTokenSource)
         {
+            // Duplicated code from SkippableFactTestCase. I'm sure we could find a way to de-dup with some thought.
             var skipMessageBus = new SkippableFactMessageBus(messageBus, constructorArguments);
             var result = await base.RunAsync(diagnosticMessageSink, skipMessageBus, constructorArguments, aggregator, cancellationTokenSource);
             if (skipMessageBus.DynamicallySkippedTestCount > 0)
