+++ Tests/CsDebugScript.Tests.Native/DebugControlTest.cs
