+++ Tests/CsDebugScript.Tests.Native/InteractiveModeTests.cs
