+++ Tests/CsDebugScript.Tests.Native/PdbReaderTests.cs
