+++ Tests/CsDebugScript.Tests.UI/CsDebugScript.Tests.UI.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
   <Import Project="..\..\SharedAssemblyInfo.props" />
 
   <PropertyGroup>
@@ -24,7 +24,7 @@
 
   <ItemGroup>
     <ProjectReference Include="$(EnlistmentRoot)\Source\CsDebugScript.UI\CsDebugScript.UI.csproj" />
-    <ProjectReference Include="$(EnlistmentRoot)\Tests\CsDebugScript.Tests\CsDebugScript.Tests.csproj" />
+    <ProjectReference Include="$(EnlistmentRoot)\Tests\CsDebugScript.Tests.Native\CsDebugScript.Tests.Native.csproj" />
   </ItemGroup>
 
   <ItemGroup>
