--- Tests/CsDebugScript.Tests/Utils/SkippableTheoryAttribute.cs
@@ -1,10 +1,0 @@
-﻿using Xunit;
-using Xunit.Sdk;
-
-namespace CsDebugScript.Tests.Utils
-{
-    [XunitTestCaseDiscoverer("CsDebugScript.Tests.Utils.SkippableTheoryDiscoverer", "CsDebugScript.Tests")]
-    public class SkippableTheoryAttribute : TheoryAttribute
-    {
-    }
-}
